+++ trendi_core/api/tools/route.py
@@ -1,3 +1,4 @@
+# global libs
 from ..constants import products_per_ip_pid
 import maxminddb
