+++ trendi_core/master_tools/imgUtils.py
@@ -4,8 +4,6 @@ from __future__ import print_function
 import requests
 from cv2 import imdecode, imwrite
 import logging
-import os
-import time
 import numpy as np
 import cv2
 import dlib
@@ -28,7 +26,7 @@ def label_img(image_or_url):
     except:
         return None
 
-# TODO - not finished yet
+
 def url_to_img_array(url):
     if not isinstance(url, basestring):
         logging.warning("input is neither an ndarray nor a string, so I don't know what to do")
@@ -37,185 +35,71 @@ def url_to_img_array(url):
     # replace_https_with_http:
     if 'http' in url and 'https' not in url:
         url = url.replace("https", "http")
-        img_url = url_or_path_to_image_file_or_cv2_image_array
-        try:
-            # print("trying remotely (url) ")
-            headers = {'User-Agent': USER_AGENT}
-            response = requests.get(img_url, headers=headers)  # download
-            img_array = imdecode(np.asarray(bytearray(response.content)), 1)
-        except requests.ConnectionError:
-            logging.warning("connection error - check url or connection")
-            return None
-        except:
-            logging.warning(" error other than connection error - check something other than connection")
-            return None
-
-# TODO - not finished yet
-def path_to_img_array(path):
-    if not isinstance(path, basestring):
-        logging.warning("input is neither an ndarray nor a string, so I don't know what to do")
+    try:
+        headers = {'User-Agent': USER_AGENT}
+        response = requests.get(url, headers=headers)
+        img_array = imdecode(np.asarray(bytearray(response.content)), 1)
+    except requests.ConnectionError:
+        logging.warning("connection error - check url or connection")
+        return None
+    except:
+        logging.warning(" error other than connection error - check something other than connection")
         return None
 
-# TODO - not finished yet
-def download_img_array(img_array, download_path):
-    filename = \
-        url_or_path_to_image_file_or_cv2_image_array.split('/')[-1].split('#')[0].split('?')[-1].split(':')[-1]
-    filename = os.path.join(download_directory, filename)
-
-    if filename.endswith('jpg') or filename.endswith('jpeg') or filename.endswith('.bmp') or \
-        filename.endswith('tiff'):
-        pass
-    else:  # there's no 'normal' filename ending so add .jpg
-        filename = filename + '.jpg'
-
-#TODO - this function need to be divided to 3
-def get_cv2_img_array(url_or_path_to_image_file_or_cv2_image_array, convert_url_to_local_filename=False, download=False,
-                      download_directory='images', replace_https_with_http=True):
-    """
-    Get a cv2 img array from a number of different possible inputs.
-    :param url_or_path_to_image_file_or_cv2_image_array:
-    :param convert_url_to_local_filename:
-    :param download:
-    :param download_directory:
-    :return: img_array
-    """
-    got_locally = False
-
-    # # first check if we already have a numpy array
-    # if isinstance(url_or_path_to_image_file_or_cv2_image_array, np.ndarray):
-    #     img_array = url_or_path_to_image_file_or_cv2_image_array
-
-    # otherwise it's probably a string, check what kind
-    elif isinstance(url_or_path_to_image_file_or_cv2_image_array, basestring):
-        # try getting url locally by changing url to standard name
-        if convert_url_to_local_filename:  # turn url into local filename and try getting it again
-            # filename = url_or_path_to_image_file_or_cv2_image_array.split('/')[-1].split('#')[0].split('?')[0]
-            # jeremy changed this since it didn't work with url -
-            # https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcR2oSMcnwErH1eqf4k8fvn2bAxvSdDSbp6voC7ijYJStL2NfX6v
-            # TODO: find a better way to create legal filename from url
-            filename = \
-                url_or_path_to_image_file_or_cv2_image_array.split('/')[-1].split('#')[0].split('?')[-1].split(':')[-1]
-            filename = os.path.join(download_directory, filename)
-
-            if filename.endswith('jpg') or filename.endswith('jpeg') or filename.endswith('.bmp') or \
-                    filename.endswith('tiff'):
-                pass
-            else:  # there's no 'normal' filename ending so add .jpg
-                filename = filename + '.jpg'
-
-            img_array = get_cv2_img_array(filename, convert_url_to_local_filename=False, download=download,
-                                          download_directory=download_directory)
-            # maybe return(get_cv2 etc) instead of img_array =
-            if img_array is not None:
-                # print('got ok array calling self locally')
-                return img_array
-            else:  # couldnt get locally so try remotely
-                # print('trying again remotely since using local filename didnt work, download=' + str( download) + ' fname:' + str(filename))
-                return (
-                    get_cv2_img_array(url_or_path_to_image_file_or_cv2_image_array, convert_url_to_local_filename=False,
-                                      download=download,
-                                      download_directory=download_directory))  # this used to be 'return'
-        # put images in local directory
-        else:
-            # get remotely if its a url, get locally if not
-            if "://" in url_or_path_to_image_file_or_cv2_image_array:
-                if replace_https_with_http:
-                    url_or_path_to_image_file_or_cv2_image_array = url_or_path_to_image_file_or_cv2_image_array.replace(
-                        "https", "http")
-                img_url = url_or_path_to_image_file_or_cv2_image_array
-                try:
-                    # print("trying remotely (url) ")
-                    headers = {'User-Agent': USER_AGENT}
-                    response = requests.get(img_url, headers=headers)  # download
-                    img_array = imdecode(np.asarray(bytearray(response.content)), 1)
-                except requests.ConnectionError:
-                    logging.warning("connection error - check url or connection")
-                    return None
-                except:
-                    logging.warning(" error other than connection error - check something other than connection")
-                    return None
-
-            else:  # get locally, since its not a url
-                # print("trying locally (not url)")
-                img_path = url_or_path_to_image_file_or_cv2_image_array
-                try:
-                    img_array = cv2.imread(img_path)
-                    if img_array is not None:
-                        # print("success trying locally (not url)")
-                        got_locally = True
-                    else:
-                        # print('couldnt get locally (in not url branch)')
-                        return None
-                except:
-                    # print("could not read locally, returning None")
-                    logging.warning("could not read locally, returning None")
-                    return None  # input isn't a basestring nor a np.ndarray....so what is it?
-    else:
-        logging.warning("input is neither an ndarray nor a string, so I don't know what to do")
-        return None
-
-    # After we're done with all the above, this should be true - final check that we're outputting a good array
-    if not (isinstance(img_array, np.ndarray) and isinstance(img_array[0][0], np.ndarray)):
-        print("Bad image - check url/path/array:" + str(
-            url_or_path_to_image_file_or_cv2_image_array) + 'try locally' + str(
-            convert_url_to_local_filename) + ' dl:' + str(
-            download) + ' dir:' + str(download_directory))
-        logging.warning("Bad image - check url/path/array:" + str(
-            url_or_path_to_image_file_or_cv2_image_array) + 'try locally' + str(
-            convert_url_to_local_filename) + ' dl:' + str(
-            download) + ' dir:' + str(download_directory))
-        return (None)
-    # if we got good image and need to save locally :
-    if download:
-        if not got_locally:  # only download if we didn't get file locally
-            if not os.path.isdir(download_directory):
-                os.makedirs(download_directory)
-            if "://" in url_or_path_to_image_file_or_cv2_image_array:  # its a url, get the bifnocho
-                if replace_https_with_http:
-                    url_or_path_to_image_file_or_cv2_image_array = url_or_path_to_image_file_or_cv2_image_array.replace(
-                        "https", "http")
-                filename = \
-                    url_or_path_to_image_file_or_cv2_image_array.split('/')[-1].split('#')[0].split('?')[-1].split(':')[
-                        -1]
-                filename = os.path.join(download_directory, filename)
-            else:  # its not a url so use straight
-                filename = os.path.join(download_directory, url_or_path_to_image_file_or_cv2_image_array)
-            if filename.endswith('jpg') or filename.endswith('jpeg') or filename.endswith('.bmp') or filename.endswith(
-                    'tiff'):
-                pass
-            else:  # there's no 'normal' filename ending
-                filename = filename + '.jpg'
-            try:  # write file then open it
-                # print('filename for local write:' + str(filename))
-                write_status = imwrite(filename, img_array)
-                max_i = 50  # wait until file is readable before continuing
-                gotfile = False
-                for i in xrange(max_i):
-                    try:
-                        with open(filename, 'rb') as _:
-                            gotfile = True
-                    except IOError:
-                        time.sleep(10)
-                if gotfile == False:
-                    print('Could not access {} after {} attempts'.format(filename, str(max_i)))
-                    raise IOError('Could not access {} after {} attempts'.format(filename, str(max_i)))
-            except:  # this is prob unneeded given the 'else' above
-                print('unexpected error in Utils calling imwrite')
     return img_array
 
 
+def path_to_img_array(img_path):
+    if not isinstance(img_path, basestring):
+        logging.warning("bad input - path string is expected")
+        return None
+
+    try:
+        img_array = cv2.imread(img_path)
+    except:
+        logging.warning("could not read locally, returning None")
+        return None
+
+    return img_array
+
+
+def download_img_array(img_array, savename):
+    if any(savename.endswith(x) for x in ['.jpg','.jpeg','.bmp','.tiff']):
+        pass
+    else:  # there's no 'normal' filename ending so add .jpg
+        savename += '.jpg'
+
+    try:
+        imwrite(savename, img_array)
+    except:  # this is prob unneeded given the 'else' above
+        print('unexpected error in Utils calling imwrite')
+        return False
+
+    return True
+
+
+def download_img_url(img_url, savename):
+    img_array = url_to_img_array(img_url)
+    if img_array is None:
+        return False
+    else:
+        return download_img_array(img_array, savename)
+
+
 def standard_resize(image, max_side):
-    original_h, original_w, _  = image.shape
+    original_h, original_w, _ = image.shape
     if all(side < max_side for side in [original_h, original_w]):
         return image, 1
     aspect_ratio = float(np.amax((original_w, original_h))/float(np.amin((original_h, original_w))))
     resize_ratio = float(float(np.amax((original_w, original_h))) / max_side)
+
     if original_w >= original_h:
         new_w = max_side
         new_h = max_side/aspect_ratio
     else:
         new_h = max_side
         new_w = max_side/aspect_ratio
+
     resized_image = cv2.resize(image, (int(new_w), int(new_h)))
     return resized_image, resize_ratio
